Fixed bug for onClick and onLongClick in InstalledApps
We were taking the app name and using that to get the package name
of a given installed app. When doing this, if two apps had the same
name (or "label"), when we got the package name, it would always get
whichever one appeared first in packageNames. Instead, we can grab
the packageName using the position that already is passed in via the
onClick and onLongClick events to always grab the right one.

I noticed we were already grabbing the packageName in this way
in the FavouriteAppsActivity file.

Should resolve #78